implemented some validations

+++ Application/Services/CourseService.cs
@@ -2,6 +2,7 @@ using Application.DTOS;
 using Application.Models;
 using Application.ServiceInterfaces;
 using AutoMapper;
+using FluentValidation;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -14,14 +15,22 @@ namespace Application.Services
     {
         private readonly IUnitOfWork _UnitOfWork;
         private readonly IMapper _mapper;
-        public CourseService(IUnitOfWork unitOfWork, IMapper mapper) 
+        private readonly IValidator<CreateCourseDTO> _validator;
+        private readonly IValidator<UpdateCourseDTO> _UpdateValidator;
+        public CourseService(IUnitOfWork unitOfWork, IMapper mapper, IValidator<CreateCourseDTO> validator, IValidator<UpdateCourseDTO> UpdateValidator) 
         {
             _UnitOfWork = unitOfWork;
             _mapper = mapper;
+            _validator = validator;
+            _UpdateValidator = UpdateValidator;
         }
 
         public async Task<bool> AssignTrainerToCourseUsingSP(int Id, int TrainerId)
         {
+            if (Id <= 0) throw new ArgumentException("Course ID must be greater than 0", nameof(Id));
+            if (TrainerId <= 0) throw new ArgumentException("Trainer ID must be greater than 0", nameof(TrainerId));
+
+
             var result = await _UnitOfWork.CourseRepository.AssignTrainerToCourseUsingSP(Id, TrainerId);
 
            await _UnitOfWork.CompleteAsync();
@@ -30,7 +39,9 @@ namespace Application.Services
         }
 
         public async Task<bool> CreateCourseUsingSP(CreateCourseDTO courseDTO)
-        { 
+        {
+            await _validator.ValidateAndThrowAsync(courseDTO);
+
             var course = _mapper.Map<Course>(courseDTO);    
 
             var result = await _UnitOfWork.CourseRepository.CreateCourseUsingSP(course);
@@ -42,6 +53,8 @@ namespace Application.Services
 
         public async Task<bool> DeleteCourseUsingSP(int id)
         {
+            if (id <= 0) throw new ArgumentException("Course ID must be greater than 0", nameof(id));
+
             var result = await _UnitOfWork.CourseRepository.DeleteCourseUsingSP(id);
 
           await  _UnitOfWork.CompleteAsync();
@@ -61,7 +74,10 @@ namespace Application.Services
 
         public async Task<bool> SetCourseCpacityUsingSP(int Capacity, int id)
         {
-           var result = await _UnitOfWork.CourseRepository.SetCourseCpacityUsingSP(Capacity, id);
+            if (id <= 0) throw new ArgumentException("Course ID must be greater than 0", nameof(id));
+            if (Capacity <= 0) throw new ArgumentException("Trainer ID must be greater than 0", nameof(Capacity));
+
+            var result = await _UnitOfWork.CourseRepository.SetCourseCpacityUsingSP(Capacity, id);
 
            await _UnitOfWork.CompleteAsync();
 
@@ -70,10 +86,19 @@ namespace Application.Services
 
         public async Task<bool> UpdateCourseUsingSP(int id, UpdateCourseDTO courseDTO)
         {
+
+
             var existingCourse = await _UnitOfWork.CourseRepository.GetByIdAsync(id);
 
             if (existingCourse == null)
                 return false;
+            if(existingCourse.TrainerId > 0)
+            {
+                throw new ArgumentException("a Trainer is already assigned to this course", nameof(existingCourse.TrainerId));
+            }
+
+            await _UpdateValidator.ValidateAndThrowAsync(courseDTO);
+
 
             _mapper.Map(courseDTO, existingCourse);
